backend/requirements.txt
fastapi==0.95.2
uvicorn==0.23.2
psycopg2-binary==2.9.7
httpx==0.28.1